+++ Non-SAS-Viya-Integrations/LLM-Call.py
@@ -4,7 +4,7 @@ import json
 import requests
 
 llm_endpoint = 'server_url/llm'
-llm_name = 'llama_31_405b'
+llm_name = 'gpt_41_mini'
 
 system_prompt = 'You are a helpful assistant that provides a short summary of the users input'
 user_prompt = 'A Boy was given permission to put his hand into a pitcher to get some filberts. But he took such a great fistful that he could not draw his hand out again. There he stood, unwilling to give up a single filbert and yet unable to get them all out at once. Vexed and disappointed he began to cry.\n\n\"My boy,\" said his mother, \"be satisfied with half the nuts you have taken and you will easily get your hand out. Then perhaps you may have some more filberts some other time.\"'